Allow the installer to correct launcher errors
Sometimes the old vanilla installer can break the json profile, which the installer can fix
But it was also replacing the jar when that happened, which is unhelpful. So now it doesn't

+++ src/main/java/org/dimdev/riftloader/Main.java
@@ -117,11 +117,13 @@ public class Main {
             Files.copy(Main.class.getResourceAsStream("/profile.json"), versionJson.toPath(), StandardCopyOption.REPLACE_EXISTING);
 
             File fakeJar = new File(minecraftFolder, "versions/1.13.2-rift-@VERSION@/1.13.2-rift-@VERSION@.jar");
-            File maybeRealJar = new File(minecraftFolder, "versions/1.13.2/1.13.2.jar");
-            if (maybeRealJar.exists()) {
-            	Files.copy(maybeRealJar.toPath(), fakeJar.toPath(), StandardCopyOption.REPLACE_EXISTING);
-            } else {
-            	fakeJar.createNewFile();
+            if (!fakeJar.exists()) {
+            	File maybeRealJar = new File(minecraftFolder, "versions/1.13.2/1.13.2.jar");
+            	if (maybeRealJar.exists()) {
+            		Files.copy(maybeRealJar.toPath(), fakeJar.toPath(), StandardCopyOption.REPLACE_EXISTING);
+                } else {
+                	fakeJar.createNewFile();
+                }
             }
 
             // Make mods directory
